Start game after canvas resize.
With the code prior to this commit, it was possible that the setup
function ran before the canvas was resized to the size of the window.
This caused the initial starting positions of the asteroids and player
to be wrong.
This fix relies on the canvas being resized to the full size of the
browser window when the page loads. If the canvas goes through multiple
resizes, this fix may not be enough.